Update kill.js
added checking against WHO field if args are present

+++ commands/kill.js
@@ -4,6 +4,7 @@ module.exports = {
 	allowedChannels: ['console', 'channel2actionnews'],
 	allowedUsers: (args, msg, groups) => process.isAdmin(msg.author.id),
 	execute(msg, args) {
+		if(args.length > 0 && args[0].toLowerCase() !== process.env.WHO.toLowerCase()) return;
 		const chan = msg.channel;
 		chan.send('Shutting down.').then(message => message.client.destroy());
 	},
